:recycle: Refactored food scan logic in designated component
